Add testing for Python 3.13

+++ test/test_execution.py
@@ -213,11 +213,22 @@ def test_signal_exit(temp_output_dir: str) -> None:
 
     thread = threading.Thread(target=runner)
     thread.start()
-    sleep(0.3)  # make sure the process is started
-    ps_command = "ps aux | grep '[s]leep 60.74016230000801'"  # brackets to not match grep process
-    ps_output = subprocess.check_output(ps_command, shell=True).decode()
-    pid = int(ps_output.split()[1])
-    os.kill(pid, signal.SIGTERM)
+    retries = 20
+    pid = None
+    for i in range(retries):
+        try:
+            ps_command = "ps auxww | grep '[s]leep 60.74016230000801'"  # brackets to not match grep process
+            ps_output = subprocess.check_output(ps_command, shell=True).decode()
+            pid = int(ps_output.split()[1])
+            break
+        except subprocess.CalledProcessError as e:
+            print(f"Attempt {i} failed with msg: {e}", file=sys.stderr)
+            sleep(0.1)  # Retry after a short delay
+
+    if pid is not None:
+        os.kill(pid, signal.SIGTERM)
+    else:
+        raise RuntimeError("Failed to find sleep process")
 
     thread.join()
     # Cannot retrieve the exit code from the thread, it is written to the file
